requirements.txt
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scipy==1.11.4
scikit-learn==1.3.2
pycryptodome==3.19.0
